Remove test staging step in deployment pipeline
Whilst preparing for the migration of PVB from template deploy to
kubernetes we planned to have a step in the deployment pipeline that
would run the integration tests against staging, once deployed, so we
could check that nothing had broken before deploying to production.
However, as PVB staff is whitelisted we have run into issues where
Circle is unable to access the site to complete the tasks within the
integration tests.  We discussed this as a team and have come to the
agreement that we will remove this step and run it manually from our own
machines, that will have access.  This is not an ideal solution, but we
do not have the capacity within the team at present to explore the
possibility of switching staging to point at T3 rather than pre-prod.
If we could achieve this in the future we would then be able to remove
the whitelisting restrictions against staging and add this step back
into the deployment pipeline.