Build maintenance page
In order for us to be able to complete the PVB migration we need to put
the service into maintenance mode, and to help us achieve this we are
making use of the Turnout gem. This allows us to serve up a
custom maintenance page in front of all our routes, and return 503s.

This PR will be reverted as part of the cleanup of the maintenance work.